// src/api/generated/models/getAuthMe200UserMembershipsItem.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { GetAuthMe200UserMembershipsItemRole } from './getAuthMe200UserMembershipsItemRole';
import type { GetAuthMe200UserMembershipsItemStatus } from './getAuthMe200UserMembershipsItemStatus';
import type { GetAuthMe200UserMembershipsItemPermissions } from './getAuthMe200UserMembershipsItemPermissions';
import type { GetAuthMe200UserMembershipsItemEffectivePermissionsItem } from './getAuthMe200UserMembershipsItemEffectivePermissionsItem';
import type { GetAuthMe200UserMembershipsItemInvitedAt } from './getAuthMe200UserMembershipsItemInvitedAt';
import type { GetAuthMe200UserMembershipsItemJoinedAt } from './getAuthMe200UserMembershipsItemJoinedAt';
import type { GetAuthMe200UserMembershipsItemCompany } from './getAuthMe200UserMembershipsItemCompany';

export type GetAuthMe200UserMembershipsItem = {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  role: GetAuthMe200UserMembershipsItemRole;
  status: GetAuthMe200UserMembershipsItemStatus;
  permissions: GetAuthMe200UserMembershipsItemPermissions;
  effectivePermissions: GetAuthMe200UserMembershipsItemEffectivePermissionsItem[];
  invitedAt: GetAuthMe200UserMembershipsItemInvitedAt;
  joinedAt: GetAuthMe200UserMembershipsItemJoinedAt;
  company?: GetAuthMe200UserMembershipsItemCompany;
};