// src/api/generated/models/partnerCompany.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { PartnerCompanyTaxId } from './partnerCompanyTaxId';
import type { PartnerCompanyCountry } from './partnerCompanyCountry';

export interface PartnerCompany {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  name: string;
  taxId: PartnerCompanyTaxId;
  country: PartnerCompanyCountry;
}